backend/pyproject.toml
[tool.poetry]
name = "shopease-backend"
version = "0.1.0"
description = "ShopEase Backend with FastAPI"
authors = ["Musfira"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.113.0"
uvicorn = "^0.30.6"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.34"
passlib = "^1.7.4"
email-validator = "^2.2.0"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
